Update Test module to get unit tests working for Elm 0.17
Due to changes in dependencies and the absorption of elm-test into
elm-community, I could remove the test runner.

In order to run the tests...
cd tests
elm-test Tests.elm

There is some funky output, but I can live with that problem for now
because it reports on the test results accurately.